Force new window on reopen project
I’m doing this to see if this fixes the problem, because I’m not 100% sure what I’m doing.

I can only trace the delegated calls back as far as the application delegate – after that I’m not sure where they go, so I’m not sure where I should fix this.